* hotfix/0.13.1: (98 commits)
  version bump
  MARS-519: fix uniqueness of filenames
  Fix timer
  Bring back Timer log to Log::info()
  ECKIT-113 Context: remove explicit template types for make_pair
  ECKIT-153 LAPACK feature no longer experimental
  ECKIT-153 ECKIT-127 geometry, linalg, maths libraries no longer experimental
  ECKIT-153 Eigen feature is no longer experimental
  ECKIT-142 Move option tests out of experimental
  ECKIT-153 Move geometry out of experimental
  ECKIT-153 Move geometry out of experimental
  Add CodeLocation to Content exceptions
  CacheLRU: constness
  ECKIT-152: hash hostname in unique
  Revert "ECKIT-112 test_floatcompare: disable floating point trapping before comparing infinity"
  Fix warnings in test_resource
  Fix warnings in mpi
  Fix warnings in BTree
  Fix warnings in UserInput
  ECKIT-112 test_floatcompare: disable floating point trapping before comparing infinity
  ...

+++ src/experimental/eckit/mpi/Collectives.h
@@ -144,8 +144,8 @@ inline int all_to_all( const Comm& comm,
 {
   int cnt;
   int mpi_size = comm.size();
-  ASSERT( sendvec.size() == mpi_size );
-  ASSERT( recvvec.size() == mpi_size );
+  ASSERT( sendvec.size() == size_t(mpi_size) );
+  ASSERT( recvvec.size() == size_t(mpi_size) );
   // Get send-information
   std::vector<int> sendcounts(mpi_size);
   std::vector<int> senddispls(mpi_size);
